fix: names of downloaded subs

+++ server/utils/subtitleDownloader.js
@@ -14,17 +14,24 @@ const download = (url, path, callback) => {
 };
 
 const decompressFile = (filepath, name) => {
+  if (!fs.existsSync(filepath)) {
+    return;
+  }
+
   const fileData = fs.readFileSync(filepath);
 
   gunzip(fileData, (err, bin) => {
     const newPath = path.resolve(__dirname, '..', 'videos', name);
 
     fs.writeFileSync(newPath, bin);
-    fs.unlinkSync(filepath);
+
+    if (fs.existsSync(filepath)) {
+      fs.unlinkSync(filepath);
+    }
   });
 };
 
-const dowloadFile = (subtitle, name) => {
+const dowloadFile = async (subtitle, name) => {
   const { SubDownloadLink, SubFileName } = subtitle;
   const newFileName = `${name}${path.extname(SubFileName)}`;
   const zipPath = path.join(__dirname, `${name}.gz`);
@@ -38,19 +45,20 @@ async function downloadSubtitleByHash(file, token) {
   const resultsBr = await subtitler.api.searchForFile(token, 'pob', file);
 
   if (resultsBr.length) {
-    resultsBr.forEach(sub => dowloadFile(sub, `${name}_${brIndex}_pob`));
-    brIndex++;
+    resultsBr.forEach(sub => {
+      dowloadFile(sub, `${name}_${brIndex}_pob`);
+      brIndex++;
+    });
   }
 
   const resultsEng = await subtitler.api.searchForFile(token, 'eng', file);
 
   if (resultsEng.length) {
-    resultsEng.forEach(sub => dowloadFile(sub, `${name}_${enIndex}_eng`));
-    enIndex++;
+    resultsEng.forEach(sub => {
+      dowloadFile(sub, `${name}_${enIndex}_eng`);
+      enIndex++;
+    });
   }
-
-  // return resultsBr.length || resultsEng.length ? true : false;
-  return false;
 }
 
 async function downloadSubtitleByTitle(file, token) {
@@ -59,41 +67,34 @@ async function downloadSubtitleByTitle(file, token) {
   const resultsBr = await subtitler.api.searchForTitle(token, 'pob', name);
 
   if (resultsBr.length) {
-    resultsBr.forEach(sub => dowloadFile(sub, `${name}_${brIndex}_pob`));
-    brIndex++;
+    resultsBr.forEach(sub => {
+      dowloadFile(sub, `${name}_${brIndex}_pob`);
+      brIndex++;
+    });
   }
 
   const resultsEng = await subtitler.api.searchForTitle(token, 'eng', name);
 
   if (resultsEng.length) {
-    resultsEng.forEach(sub => dowloadFile(sub, `${name}_${enIndex}_eng`));
-    enIndex++;
+    resultsEng.forEach(sub => {
+      dowloadFile(sub, `${name}_${enIndex}_eng`);
+      enIndex++;
+    });
   }
-
-  // return resultsBr.length || resultsEng.length ? true : false;
-  return false;
 }
 
 async function downloadSubtitle(file) {
+  brIndex = 0;
+  enIndex = 0;
+
   const name = path.basename(file, path.extname(file));
 
   console.log(`>>> Download subtitle ${name}`);
 
   const token = await subtitler.api.login();
 
-  if (await downloadSubtitleByHash(file, token)) {
-    console.log(`>>> Subtitle downloaded ${name}`);
-
-    await subtitler.api.logout();
-    return;
-  }
-
-  if (await downloadSubtitleByTitle(file, token)) {
-    console.log(`>>> Subtitle downloaded ${name}`);
-
-    await subtitler.api.logout();
-    return;
-  }
+  await downloadSubtitleByHash(file, token);
+  await downloadSubtitleByTitle(file, token);
 
   await subtitler.api.logout();
 }
